added inital doc folder
