+++ docs/Introduction.rst
@@ -1,0 +1,6 @@
+Introdution
+===================
+
+Welcome to the documentation of Secure-privacy-preserving-and-federated-machine-learning !
+
+The code is available at https://github.com/XavierLessage/Secure-privacy-preserving-and-federated-machine-learning.
