+++ docs/api.rst
@@ -1,0 +1,10 @@
+API
+=====================
+
+Welcome to the API.
+
+.. autosummary::
+   :toctree: DIRNAME
+
+   DICOM_PNGJSON_Converter.DICOM_converter
+   sphinx.util.relative_uri
