+++ docs/index.rst
@@ -1,0 +1,11 @@
+Hello I'm a title
+==========================
+
+Hello I'm an API
+
+.. toctree::
+   :maxdepth: 2
+   :caption: Getting Started
+
+   Introduction
+   api
